requirements.txt
streamlit==1.32.0
pandas==2.2.1
boto3==1.34.34
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
openai==1.12.0
beautifulsoup4==4.12.3
requests==2.31.0
python-dateutil==2.8.2
qdrant-client==1.7.0 